StudentCollaborationDetailController moved to new folder Student 	renamed:    app/Http/Controllers/api/StudentCollaborationDetailController.php -> app/Http/Controllers/api/Student/StudentCollaborationDetailController.php 	modified:   routes/api/v1.php 	modified:   tests/Feature/Student/StudentCollaborationDetailControllerTest.php
