+++ tests/Feature/Student/StudentCollaborationDetailControllerTest.php
@@ -8,7 +8,7 @@ use Tests\TestCase;
 use App\Models\Student;
 use App\Models\Collaboration;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
-use App\Http\Controllers\api\StudentCollaborationDetailController;
+use App\Http\Controllers\api\Student\StudentCollaborationDetailController;
 use App\Service\StudentCollaborationDetailService;
 
 class StudentCollaborationDetailControllerTest extends TestCase
